Add src/summstats/garud.cc and move initialization of GarudStats class to that file.
